refactor: changes grid into integrationsSection
